Feature/authenticate be

+++ services/user-service/src/entry-points/api/routes.ts
@@ -1,5 +1,9 @@
 import express from "express";
 
+import {
+  assertIsAuthenticated,
+  assertIsSelfOrAdmin,
+} from "../../commons/auth/authenticator";
 import * as userUseCase from "../../domain/user-use-case";
 
 import { validateAddUserInput, validateUpdateUserInput } from "./validators";
@@ -7,9 +11,9 @@ import { validateAddUserInput, validateUpdateUserInput } from "./validators";
 export default function defineRoutes(expressApp: express.Application) {
   const router = express.Router();
 
-  router.post("/", validateAddUserInput, async (req, res, next) => {
+  router.post("/", [validateAddUserInput], async (req, res, next) => {
     try {
-      // ✅ Best Practice: Using the 3-tier architecture, routes/controller are kept thin, logic is encapsulated in a dedicated domain folder
+      await assertIsAuthenticated(req);
       const addUserResponse = await userUseCase.addUser(req.body);
       res.json(addUserResponse);
     } catch (error) {
@@ -41,8 +45,9 @@ export default function defineRoutes(expressApp: express.Application) {
   });
 
   // update user by id
-  router.put("/:id", validateUpdateUserInput, async (req, res, next) => {
+  router.put("/:id", [validateUpdateUserInput], async (req, res, next) => {
     try {
+      await assertIsSelfOrAdmin(req, parseInt(req.params.id, 10));
       const response = await userUseCase.updateUser(
         parseInt(req.params.id, 10),
         req.body,
@@ -56,9 +61,11 @@ export default function defineRoutes(expressApp: express.Application) {
   // delete user by id
   router.delete("/:id", async (req, res, next) => {
     try {
+      await assertIsSelfOrAdmin(req, parseInt(req.params.id, 10));
       await userUseCase.deleteUser(parseInt(req.params.id, 10));
       res.status(200).end();
     } catch (error) {
+      console.log(error);
       next(error);
     }
   });
@@ -67,7 +74,7 @@ export default function defineRoutes(expressApp: express.Application) {
 
   expressApp.get("/health", async (req, res, next) => {
     try {
-      res.status(200).end("Healthy");
+      res.status(200).send("healthy");
     } catch (error) {
       next(error);
     }
